Optimize key encoding
